+++ ft_printf.c
@@ -1,4 +1,16 @@
-#include "printf.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_printf.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: msharifi <msharifi@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/06 15:35:56 by msharifi          #+#    #+#             */
+/*   Updated: 2022/05/06 15:35:56 by msharifi         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "ft_printf.h"
 
 const char	*ft_search_arg(va_list arg, const char *format, t_lw *sc)
 {
@@ -11,14 +23,14 @@ const char	*ft_search_arg(va_list arg, const char *format, t_lw *sc)
 	else if (*format == 'u')
 		sc->len += ft_putnbr_unsigned(va_arg(arg, unsigned int));
 	else if (*format == 'x')
-		sc->len += ft_puthex(va_arg(arg, int), "0123456789abcdef");
+		sc->len += ft_puthex(va_arg(arg, unsigned int), "0123456789abcdef");
 	else if (*format == 'X')
-		sc->len += ft_puthex(va_arg(arg, int), "0123456789ABCDEF");
+		sc->len += ft_puthex(va_arg(arg, unsigned int), "0123456789ABCDEF");
 	else if (*format == 'p')
-		sc->len += ft_putaddr(va_arg(arg, int));
+		sc->len += ft_putaddr(va_arg(arg, unsigned long int));
 	else if (*format == '%')
 		sc->len += ft_putchar('%');
-	format++; // Avancer de 1 apres avoir lu le %[.] 
+	format++;
 	return (format);
 }
 
@@ -26,10 +38,10 @@ const char	*ft_read_text(t_lw *sc, const char *format)
 {
 	char	*next;
 
-	next = ft_strrchr(format); // until the next "%"
+	next = ft_strrchr(format);
 	if (next)
 		sc->width = next - format;
-	else 
+	else
 		sc->width = ft_strlen(format);
 	write(1, format, sc->width);
 	sc->len += sc->width;
@@ -41,8 +53,8 @@ const char	*ft_read_text(t_lw *sc, const char *format)
 int	ft_printf(const char *format, ...)
 {
 	t_lw	sc;
-
 	va_list	arg;
+
 	va_start(arg, format);
 	sc.len = 0;
 	sc.width = 0;
